Add 4 files and 2 folders to my final project
https://nominatim.org/release-docs/latest/api/Overview
https://developer.yr.no/doc/locationforecast/HowTO/

The folder Nominatim uses the Open street map APIs nominatim reverse and
nominatim search. The folder Yr_no simply uses the Location Forecast API
from Yr.no the Norwegian meteorological institute. yrno_modified.py is
the same as yrno.py so far. The former was created to be modified to
include a database that saves the forecast and updates it.