Massive updates of both acinn_parser and yrno_modified
These updates allow to run first the acinn_parser on the acinn website
and retrieve the lat long for each weather station (then stored in a new
SQLite3 database). Then running the yrno_modified.py program it is possible to predict weather for the next
55 hours at each of these stations and save this new data on a new table inside the same SQLite3 database.

As of now the table updates every time the code is ran and there is no
information about  how many days in advance the forecast is made.