Visualization based on the D3.js library and gline
Three files were added to the project inside the Yr_no folder.
gline2.py is a python library which is meant to visualize the data using
the javascript d3 library and saving a gline.js file. Then gline.htm is
opened onto the browser and  uses this gline.js file for visualization